refactor: hold response rendering apart from the runtime
deepagents_runtime.py was 6,891 lines. Every fix this week meant reading past
several hundred lines of string building to reach the logic underneath it, and
the mistakes made in that file were caught late.

Twenty-four rendering functions and the seven evidence-prefix constants they
use move to response_format.py. They were chosen because they are closed under
themselves: they call nothing but each other, touch no state, make no retrieval
or policy decision, and take no branch that depends on the runtime. They turn
data that has already been decided into the exact strings the model and the
shopper read.

That is also why they are worth holding apart. Their output is a contract --
several are asserted byte-for-byte by the evidence tests -- so keeping them in
one place makes a wording change a visible diff rather than a quietly different
string.

Nothing was edited during the move. Every one of the 31 moved symbols is
byte-identical to its previous version, checked by comparing parsed source
against the merge base rather than by inspection. The runtime imports the names
it still calls; eight turned out to be used only by the moved code and are now
internal to the new module.

No behaviour change is intended and none is expected. 982 tests pass unchanged.
